requirements.txt
numpy==2.2.3
python-dotenv==1.0.1
python-telegram-bot==21.11.1
torch==2.5.1
transformers==4.48.3
